+++ R/golem.R
@@ -26,7 +26,7 @@
 #' # move to package
 #' setwd(tmp)
 #' 
-#' # scaffold ambiorix
+#' # scaffold golem
 #' scaffold_golem()
 #' 
 #' # clean up