+++ inst/bare/index.js
@@ -1,0 +1,7 @@
+import 'shiny';
+
+// In shiny server use:
+// session$sendCustomMessage('packer-alert', 'hello packer!')
+Shiny.addCustomMessageHandler('packer-alert', (msg) => {
+  alert(msg)
+});